fix: tailwind also uses the correct directory
