+++ packages/lib/plugin-connector/src/plugin-discovery/resolve-evidence-config.js
@@ -1,34 +1,6 @@
 import { isValidPackage } from './is-valid-package';
-import path from 'path';
 import { loadConfig } from './load-config';
-import { createRequire } from 'module';
-
-/**
- *
- * @param {string} packageName
- * @returns {string}
- */
-const findPluginDirectory = (packageName) => {
-	// TODO: it would probably be better to use findPackageJSON but only available in node ^23
-	//       https://nodejs.org/api/module.html#modulefindpackagejsonspecifier-base
-	const mainPackageExportFile = createRequire(import.meta.url).resolve(packageName);
-
-	let packageRoot = mainPackageExportFile;
-
-	const packageNamePath = packageName.replace('/', path.sep);
-
-	while (packageRoot !== path.sep && packageRoot.endsWith(packageNamePath) === false) {
-		packageRoot = path.dirname(packageRoot);
-	}
-
-	if (packageRoot.endsWith(packageName) === false) {
-		throw new Error(
-			`failed to find plugin package root for ${packageName}, export file ${mainPackageExportFile}`
-		);
-	}
-
-	return packageRoot;
-};
+import findPluginDirectory from './find-plugin-directory.js';
 
 /**
  * Validates that the given package name exists and is a valid plugin package
